+++ packages/shared/src/index.ts
@@ -1,3 +1,4 @@
+export * from "./cloudDevEnvironments";
 export * from "./cursorApi";
 export * from "./usageAnalytics";
 export * from "./usageChart";